+++ token/program-2022/src/instruction.rs
@@ -29,8 +29,6 @@ pub enum TokenInstruction {
     /// Otherwise another party can acquire ownership of the uninitialized
     /// account.
     ///
-    /// Any extensions must be initialized before this instruction.
-    ///
     /// Accounts expected by this instruction:
     ///
     ///   0. `[writable]` The mint to initialize.
@@ -55,8 +53,6 @@ pub enum TokenInstruction {
     /// `CreateAccount` instruction that creates the account being initialized.
     /// Otherwise another party can acquire ownership of the uninitialized
     /// account.
-    ///
-    /// All required extensions will be initialized at the same time.
     ///
     /// Accounts expected by this instruction:
     ///
@@ -97,8 +93,8 @@ pub enum TokenInstruction {
     /// amounts of SOL and Tokens will be transferred to the destination
     /// account.
     ///
-    /// If the accounts contain `AccountTransferFeeExtension`s, this will fail.  Mints with
-    /// the `MintTransferFeeExtension` are required in order to assess the fee.
+    /// If either account contains an `AccountTransferFee` extension, this will fail.
+    /// Mints with the `MintTransferFee` extension are required in order to assess the fee.
     ///
     /// Accounts expected by this instruction:
     ///
@@ -112,6 +108,10 @@ pub enum TokenInstruction {
     ///   1. `[writable]` The destination account.
     ///   2. `[]` The source account's multisignature owner/delegate.
     ///   3. ..3+M `[signer]` M signer accounts.
+    #[deprecated(
+        since = "4.0.0",
+        note = "please use `TransferChecked` or `TransferCheckedWithFee` instead"
+    )]
     Transfer {
         /// The amount of tokens to transfer.
         amount: u64,
@@ -207,7 +207,7 @@ pub enum TokenInstruction {
     /// Close an account by transferring all its SOL to the destination account.
     /// Non-native accounts may only be closed if its token amount is zero.
     ///
-    /// Accounts with the `AccountTransferFeeExtension` may only be closed if the withheld
+    /// Accounts with the `AccountTransferFee` extension may only be closed if the withheld
     /// amount is zero.
     ///
     /// Mints may be closed if they have the `MintCloseAuthority` extension and their token
@@ -267,8 +267,8 @@ pub enum TokenInstruction {
     /// decimals value is checked by the caller.  This may be useful when
     /// creating transactions offline or within a hardware wallet.
     ///
-    /// If the accounts contain `AccountTransferFeeExtension`s, the fee is withheld in the
-    /// destination account.
+    /// If either account contains an `AccountTransferFee` extension, the fee is
+    /// withheld in the destination account.
     ///
     /// Accounts expected by this instruction:
     ///
@@ -373,8 +373,6 @@ pub enum TokenInstruction {
     /// Cross Program Invocation from an instruction that does not need the owner's
     /// `AccountInfo` otherwise.
     ///
-    /// All required extensions will be initialized at the same time.
-    ///
     /// Accounts expected by this instruction:
     ///
     ///   0. `[writable]`  The account to initialize.
@@ -395,8 +393,6 @@ pub enum TokenInstruction {
     ///   0. `[writable]`  The native token account to sync with its underlying lamports.
     SyncNative,
     /// Like InitializeAccount2, but does not require the Rent sysvar to be provided
-    ///
-    /// All required extensions will be initialized at the same time.
     ///
     /// Accounts expected by this instruction:
     ///
@@ -419,8 +415,6 @@ pub enum TokenInstruction {
         m: u8,
     },
     /// Like InitializeMint, but does not require the Rent sysvar to be provided
-    ///
-    /// Any extensions must be initialized before this instruction.
     ///
     /// Accounts expected by this instruction:
     ///
@@ -451,14 +445,12 @@ pub enum TokenInstruction {
     ///
     /// The mint must have exactly enough space allocated for the base mint (82
     /// bytes), plus 83 bytes of padding, 1 byte reserved for the account type,
-    /// then space required for the extension.
+    /// then space required for this extension, plus any others.
     ///
     /// Accounts expected by this instruction:
     ///
     ///   0. `[writable]` The mint to initialize.
-    InitializeMintCloseAuthority {
-        close_authority: COption<Pubkey>,
-    },
+    InitializeMintCloseAuthority { close_authority: COption<Pubkey> },
     /// Initialize the transfer fee on a new mint.
     ///
     /// Fails if the mint has already been initialized, so must be called before
@@ -466,7 +458,7 @@ pub enum TokenInstruction {
     ///
     /// The mint must have exactly enough space allocated for the base mint (82
     /// bytes), plus 83 bytes of padding, 1 byte reserved for the account type,
-    /// then space required for the extension.
+    /// then space required for this extension, plus any others.
     ///
     /// Accounts expected by this instruction:
     ///
@@ -474,7 +466,7 @@ pub enum TokenInstruction {
     InitializeMintTransferFee {
         /// Pubkey that may update the fees
         fee_config_authority: COption<Pubkey>,
-        /// Harvest instructions must be signed by this key
+        /// Withdraw instructions must be signed by this key
         fee_withdraw_authority: COption<Pubkey>,
         /// Amount of transfer collected as fees, expressed as basis points of the
         /// transfer amount
@@ -487,9 +479,9 @@ pub enum TokenInstruction {
     /// Accounts expected by this instruction:
     ///
     ///   * Single owner/delegate
-    ///   0. `[writable]` The source account. Must include the `AccountTransferFeeExtension`.
-    ///   1. `[]` The token mint. Must include the `AccountMintFeeExtension`.
-    ///   2. `[writable]` The destination account. Must include the `AccountTransferFeeExtension`.
+    ///   0. `[writable]` The source account. Must include the `AccountTransferFee` extension.
+    ///   1. `[]` The token mint. Must include the `MintTransferFee` extension.
+    ///   2. `[writable]` The destination account. Must include the `AccountTransferFee` extension.
     ///   3. `[signer]` The source account's owner/delegate.
     ///
     ///   * Multisignature owner/delegate
@@ -513,8 +505,8 @@ pub enum TokenInstruction {
     /// Accounts expected by this instruction:
     ///
     ///   * Single owner/delegate
-    ///   0. `[writable]` The token mint. Must include the `MintTransferFeeExtension`.
-    ///   1. `[writable]` The fee receiver account. Must include the `AccountTransferFeeExtension`.
+    ///   0. `[writable]` The token mint. Must include the `MintTransferFee` extension.
+    ///   1. `[writable]` The fee receiver account. Must include the `AccountTransferFee` extension
     ///      associated with the provided mint.
     ///   2. `[signer]` The mint's `fee_withdraw_authority`
     ///
@@ -523,12 +515,12 @@ pub enum TokenInstruction {
     ///   1. `[writable]` The destination account.
     ///   2. `[]` The mint's `fee_withdraw_authority`'s multisignature owner/delegate.
     ///   3. ..3+M `[signer]` M signer accounts.
-    HarvestWithheldTokensFromMint,
+    WithdrawWithheldTokensFromMint,
     /// Permissionless instruction to transfer all withheld tokens to the mint.
     ///
     /// Succeeds for frozen accounts.
     ///
-    /// Accounts provided should include the `AccountTransferFeeExtension`. If not,
+    /// Accounts provided should include the `AccountTransferFee` extension. If not,
     /// the account is skipped.
     ///
     /// Accounts expected by this instruction:
@@ -536,7 +528,7 @@ pub enum TokenInstruction {
     ///   0. `[writable]` The mint.
     ///   1. ..1+N `[writable]` The source accounts to harvest from.
     HarvestWithheldTokensToMint,
-    /// Set transfer fee. Only supported for mints that include the `MintTransferFeeExtension`.
+    /// Set transfer fee. Only supported for mints that include the `MintTransferFee` extension.
     ///
     /// Accounts expected by this instruction:
     ///
